Removed concept of PreviousStreamRevision from CommitAttempt and replaced it with StreamRevision.
The PreviousStreamRevision property was a redundancy to the PreviousCommitSequence property in that they were both facilitating optimistic concurrency.
Only one was necessary and since PreviousCommitSequence is our primary concurrency control, it may PreviousStreamRevision obsolete.

At the same time, when attempting to commit, we need to know what the stream revision should be, which used to be PreviousStreamRevision + Events.Count.
We therefore added a property to allow us to know what revision things should be.

Lastly, some tests were added to ensure that the number commits never exceeds a stream's revision.