Made *massive* BREAKING CHANGE to ProtocolBufferSerializer--now using 16-byte key instead of 4-byte key to identify types.
According to the MSDN docs, GetHashCode() on a string is subject to change between processor architectures and versions of the .NET runtime.  This means that it's not suitable value to be used for longterm persistence such as with serialization to a disk.  We are now using MD5 serialization of a string to get a 16-byte key (which we treat as a GUID for simplicity).  This 16-byte key allows us to unambiguously identify message serialized to the disk but at a slightly higher cost in terms of storage space (16 bytes vs 4 bytes).