Added behavior and tests to make getting a stream from a snapshot forward is easier and significantly safer.
The previous commits with the new model had a fundamental issue whereby a user could query for all events and revisions after a snapshot and not find any.  This would cause the stream to be empty.  As a result, the empty stream wouldn't be able to determine what revision or commit sequence it was.  This made committing the stream impossible.  The fix was to expose an additional API method so that callers that had previously retrieved a snapshot could get everything from the revision of the snapshot forward.  This would enable the stream to have just enough information to know its revision and commit sequence.