Moved the responsibility of querying for the latest commits after a concurrency exception is raised back to the OptimisticEventStream.
Doing this has the effect of simplifying a lot of code because there is exactly one place that will query for new events after a concurrency exception is detected.  Furthermore, the stream can only query through the underlying storage reference, which is actually a reference to the OptimisticEventStore.  As such, the OptimisticEventStore will be able to cache the StreamHead properly as new commits are discovered and it can even implement additional caching of the actual commits themselves (as a performance optimization).  It also removes the awkward "Commits" property on the concurrency exception.

Lastly, this simplifies the OptimisticEventStore and Persistence implementations because they no longer have to do any additional work--the OptimisticEventStore doesn't need all kinds of extra logic to query for new commits on a concurrency exception.