Fixed issue #79 where a race condition was found when persisting commits simultaneously between multiple writers.
Now when a multiple writers attempt to insert into the commits table simultaneously, one will throw an exception which will be detected as a "UniqueKeyViolationException".  The SqlPersistenceEngine will detect the unique key violation and then re-query the table to determine if the commit is, in fact, a true duplicate (with the same commit ID) *or* if another writer has written ahead of this writer causing a concurrency violation.