Fixes #292 [Oracle] Payload of > 32KB results in exception
Because the PersistCommit oracle stament is now
in an anonymous block (to support the RETURING clause) the Payload is limited to 32KB.

In order fix this we need to use use Oracle specific concrecte types and not the ado.net / general
sql abstractions. What I did not want to do was for NEventStore to have a dependency on odp.net.managed,
nor did I want to have a seperate Oracle project and nuget packages. Instead, I am using reflection for oracle
specific operation. I *think* the structure of Oracle.ManagedDataAccess should be stable enough
(famous last words).