Adds typed parameter support and fixes performance problem #326
The @BucketId and @StreamId parameters were previously passed in as NVARCHAR values via ADO.NET, whereas the indexes are defined on VARCHAR and CHAR values. This caused (at least) SQL Server to skip the indexes, e.g., when calling OpenStream. This is now fixed by declaring the @BucketId and @StreamId parameters as "AnsiString".