Persisting null event throws #434
When you try to add an Event that is null or that
has a null body, the old code simply ignore the event.

The new code will throw because adding a null event or
an event with a null body is an error.

Updated version.MD to specify breaking changes in adding
null event semantic (now throws instead of silently ignore)

closes #434