Refs: #468 Removed virtual functions in the logging class
to improve performances in the library while logging. The IsEnabled()
properties now need to e checked before each logging operation; a bool
check is less expensive than invoking the logging function when not
needed.
This logging class is not intended to be a full fledged logging
framework and should be used internally NEventStore only.